Avoid zero-padding prover instance commitments
Upstreamed from https://github.com/zakura-core/libraries. Reproduce with:

    git format-patch -1 0ec2905 \
      -- halo2_proofs/ ':(exclude)halo2_proofs/CHANGELOG.md' \
      ':(exclude)halo2_proofs/Cargo.toml'

No --relative: halo2 is a workspace and halo2_proofs/ exists upstream.

The source commit also changed CHANGELOG.md, which the exclusions drop;
it is reapplied in its own commit at the end of the series.

Applied with git am --3way; no hunk was re-sited.

Source-Commit: 0ec2905ca3f295363abc4e1ad54bccf0a1a17e8b
